refactor: 요청 형식을 통일하기 위한 Location 분리
