+++ backend/src/main/java/hyfive/gachita/api/geocode/dto/LatLng.java
@@ -1,9 +1,3 @@
 package hyfive.gachita.api.geocode.dto;
 
-public record LatLng(double lat, double lng) {
-
-    @Override
-    public String toString() {
-        return String.format("%.15f,%.15f", lat, lng);
-    }
-}
+public record LatLng(double lat, double lng) {}
